feat(lib): extract lib/audit.php (ADR-004, sub of #907) — absorbs #912
Move the audit-log layer out of lib.php into a focused lib/audit.php
module per ADR-004 Phase 4 Task 4.2: the AUDIT_FILTER_PREFIXES const,
audit_filter_validate_prefix/_action, audit, audit_export, and
prune_audit_log. Functions stay in the global namespace (ADR-004
Option E). None read `global $config;`, so no ADR-003 conversion was
needed. The module is required from both lib.php (after lib/db.php, so
direct lib.php bootstraps see it) and init.php (after lib/db.php).

#912: prune_audit_log() carried a 3-arm switch on the driver name, each
arm implementing its own append-only-trigger bypass strategy plus an
error-path restoration arm. That logic moves into the Dialect interface
as with_append_only_bypass(PDO, callable): each concrete dialect sets up
its bypass, runs the work, and tears it down — including on the error
path. SQLite drops/recreates the two RAISE(ABORT) triggers inside a
BEGIN IMMEDIATE transaction, recreating them from its own
append_only_trigger('audit_log') output (no lib/db.php dependency);
MySQL toggles the @ipam_bypass_append_only session variable; Postgres
uses SET LOCAL on a transaction it owns only when not already in one.
prune_audit_log() collapses to a cutoff build plus a DELETE closure
handed to the dialect. Observable behaviour is byte-identical:
retentionDays <= 0 -> 0; failure -> error_log + return 0.

Adds tests/Unit/AuditExtractionParityTest.php to assert the functions
relocated to lib/audit.php and to behaviourally cover the two pure
filter validators.

Gate: php -l clean on all changed files; phpunit 1256 tests pass;
phpstan L9 no errors; phpcs clean; lib-module-linter clean. 3-driver
Playwright smoke deferred to Phase 12 per the branch pattern.

Closes #912.
