+++ Simple-PHP-IPAM/lib.php
@@ -5,6 +5,7 @@ require_once __DIR__ . '/lib/utils.php';
 require_once __DIR__ . '/lib/ip.php';
 require_once __DIR__ . '/lib/config.php';
 require_once __DIR__ . '/lib/db.php';
+require_once __DIR__ . '/lib/audit.php';
 require_once __DIR__ . '/lib/BackupClientInterface.php';
 require_once __DIR__ . '/lib/S3Client.php';
 require_once __DIR__ . '/lib/SftpClient.php';
@@ -723,94 +724,6 @@ function flash_get(): ?array
     $msg  = is_string($flash['msg']  ?? null) ? $flash['msg']  : '';
     $type = is_string($flash['type'] ?? null) ? $flash['type'] : 'info';
     return ['msg' => $msg, 'type' => $type];
-}
-
-/**
- * Allowed audit-action prefixes (categories).
- * Shared between audit.php (UI filter) and api.php (/audit endpoint).
- */
-const AUDIT_FILTER_PREFIXES = [
-    'address', 'aggregate', 'alert', 'apikey', 'audit', 'auth', 'backup',
-    'backup_run', 'config', 'contact', 'custom_field', 'db', 'destination',
-    'device', 'device_interface', 'dhcp_pool', 'export', 'import', 'mail',
-    'mfa', 'pd_pool', 'remote_backup', 'restore', 'scan', 'setting',
-    'settings', 'site', 'subnet', 'tag', 'user', 'vault', 'vlan', 'vrf',
-    'webhook',
-];
-
-/**
- * Validate an audit-prefix filter string. Returns the prefix if it matches the
- * allowlist, or '' if not. Use for ?prefix=foo style filters.
- */
-function audit_filter_validate_prefix(string $raw): string
-{
-    $p = trim($raw);
-    return ($p !== '' && in_array($p, AUDIT_FILTER_PREFIXES, true)) ? $p : '';
-}
-
-/**
- * Validate an exact audit-action filter string. Returns the action if it
- * matches the <prefix>.<verb> regex, or '' if not. Use for ?action=auth.login
- * style filters; no SQL-injection surface beyond bind.
- */
-function audit_filter_validate_action(string $raw): string
-{
-    $a = trim($raw);
-    // CR #1100: allow multi-segment actions like 'mfa.otp.fail' or
-    // 'backup.vault_key.revealed'. The previous single-dot regex
-    // rejected every audit row this codebase emits with two-dot
-    // hierarchies, so ?action=<that> filters could never match.
-    return ($a !== '' && preg_match('/^[a-z_]+(?:\.[a-z_]+)+$/', $a)) ? $a : '';
-}
-
-/**
- * Append a row to audit_log. Returns true on success, false on PDO failure
- * (with the failure logged via error_log). Callers that need to surface
- * audit failures (e.g. cron jobs) should check the return value; user-facing
- * pages can continue to ignore it. Pre-v3.26.0 this function was `void` and
- * would let exceptions propagate up the page, sometimes bubbling past the
- * page layout and rendering a blank screen.
- *
- * v3.26.0 (#1100 CR review): when invoked inside an active transaction,
- * a PDO failure RETHROWS rather than silently returning false. Callers
- * like ipam_setting_set() and auto_reserve_subnet_ips() rely on the
- * audit row landing atomically with the persisted state change; if the
- * audit insert fails, the surrounding transaction must roll back so
- * we don't commit state without its corresponding audit entry. Outside
- * an active transaction (most user-facing pages and cron tasks) the
- * legacy false-return behaviour is preserved.
- */
-function audit(PDO $db, string $action, string $entityType, ?int $entityId, string $details = ''): bool
-{
-    $u = current_user();
-    try {
-        $st = $db->prepare("INSERT INTO audit_log (user_id, username, action, entity_type, entity_id, ip, user_agent, details)
-                            VALUES (:uid,:un,:ac,:et,:eid,:ip,:ua,:dt)");
-        $st->execute([
-            ':uid' => $u['id'] ?: null,
-            ':un'  => $u['username'] ?: null,
-            ':ac'  => $action,
-            ':et'  => $entityType,
-            ':eid' => $entityId,
-            ':ip'  => client_ip() ?: null,
-            ':ua'  => to_str($_SERVER['HTTP_USER_AGENT'] ?? ''),
-            ':dt'  => $details,
-        ]);
-        return true;
-    } catch (\PDOException $e) {
-        error_log("audit failed: action={$action} entity={$entityType} id="
-            . ($entityId === null ? 'NULL' : (string)$entityId)
-            . ' err=' . $e->getMessage());
-        if ($db->inTransaction()) {
-            throw $e;
-        }
-        return false;
-    }
-}
-
-function audit_export(PDO $db, string $what, string $details = ''): void
-{
-    audit($db, "export.$what", 'system', null, $details);
 }
 
 /**
@@ -3086,108 +2999,6 @@ function housekeeping_mark_ran(): void
 
     @file_put_contents($path, json_encode(['last_run' => time()], JSON_PRETTY_PRINT));
     @chmod($path, 0600);
-}
-
-function prune_audit_log(PDO $db, int $retentionDays): int
-{
-    if ($retentionDays <= 0) return 0;
-    $cutoff = date('Y-m-d H:i:s', (int)strtotime("-{$retentionDays} days"));
-
-    // Retention routine must DELETE rows without ever leaving the
-    // append-only guarantee observable-violated. v2.10.0 #502 post-review
-    // fix: previously this dropped triggers, DELETEd, then recreated the
-    // triggers — which exposed a race window where other connections
-    // could UPDATE/DELETE audit_log. Two per-engine strategies now close
-    // the window without needing a drop/recreate cycle:
-    //
-    //   SQLite  — wrap the DELETE in BEGIN IMMEDIATE / COMMIT. SQLite DDL
-    //             is transactional AND BEGIN IMMEDIATE holds a reserved
-    //             write lock the entire time, so no other writer can
-    //             touch audit_log. The SQLite RAISE(ABORT) triggers DO
-    //             fire on DELETE, so we still have to drop+recreate on
-    //             SQLite — but now inside the reserved-lock transaction,
-    //             which is atomic from every other connection's point of
-    //             view.
-    //
-    //   MySQL   — set @ipam_bypass_append_only = 1 at session scope. The
-    //             MysqlDialect trigger bodies wrap SIGNAL in an IF block
-    //             gated on this variable (v2.10.0 #502 change to
-    //             MysqlDialect::append_only_trigger). Session variables
-    //             are per-connection so the bypass never leaks to other
-    //             connections — their SIGNAL continues to fire
-    //             unconditionally. DELETE proceeds on this connection,
-    //             every other write is still blocked. No drop/recreate
-    //             needed; the triggers stay in place the entire time.
-    //
-    //   Postgres (v2.11.0 #388) — SET LOCAL ipam.bypass_append_only = '1'
-    //             inside a transaction. The custom GUC is checked by the
-    //             PL/pgSQL trigger function (current_setting('ipam.bypass_
-    //             append_only', true) IS DISTINCT FROM '1'). SET LOCAL is
-    //             per-transaction so the bypass automatically unsets on
-    //             COMMIT/ROLLBACK — no explicit cleanup needed and no leak
-    //             risk to other connections or subsequent transactions.
-    //
-    // Error paths always attempt to restore a safe state: clear the
-    // MySQL session variable, recreate SQLite triggers via
-    // ensure_audit_log_table() as a fallback.
-    $driver = ipam_dialect()->driver_name();
-    try {
-        if ($driver === 'sqlite') {
-            $db->exec("BEGIN IMMEDIATE");
-            $db->exec("DROP TRIGGER IF EXISTS audit_log_no_update");
-            $db->exec("DROP TRIGGER IF EXISTS audit_log_no_delete");
-
-            $st = $db->prepare("DELETE FROM audit_log WHERE created_at < :cutoff");
-            $st->execute([':cutoff' => $cutoff]);
-            $pruned = $st->rowCount();
-
-            // Recreate the append-only triggers explicitly. The probe in
-            // ensure_audit_log_table() short-circuits when the table is
-            // present (it is — we only dropped the triggers, not the
-            // table), so we cannot rely on it to put the triggers back.
-            ensure_audit_log_triggers($db);
-            $db->exec("COMMIT");
-        } elseif ($driver === 'mysql') {
-            $db->exec("SET @ipam_bypass_append_only = 1");
-
-            $st = $db->prepare("DELETE FROM audit_log WHERE created_at < :cutoff");
-            $st->execute([':cutoff' => $cutoff]);
-            $pruned = $st->rowCount();
-
-            $db->exec("SET @ipam_bypass_append_only = NULL");
-        } elseif ($driver === 'pgsql') {
-            $ownTx = !$db->inTransaction();
-            if ($ownTx) {
-                $db->beginTransaction();
-            }
-            $db->exec("SET LOCAL ipam.bypass_append_only = '1'");
-
-            $st = $db->prepare("DELETE FROM audit_log WHERE created_at < :cutoff");
-            $st->execute([':cutoff' => $cutoff]);
-            $pruned = $st->rowCount();
-
-            if ($ownTx) {
-                $db->commit();
-            }
-        } else {
-            throw new \RuntimeException("prune_audit_log: unsupported driver '$driver'");
-        }
-
-        return $pruned;
-    } catch (Throwable $e) {
-        if ($driver === 'sqlite') {
-            try { $db->exec("ROLLBACK"); } catch (Throwable) {}
-            try { ensure_audit_log_table($db); } catch (Throwable) {}
-        } elseif ($driver === 'mysql') {
-            try { $db->exec("SET @ipam_bypass_append_only = NULL"); } catch (Throwable) {}
-        } elseif ($driver === 'pgsql') {
-            if (isset($ownTx) && $ownTx && $db->inTransaction()) {
-                try { $db->rollBack(); } catch (Throwable) {}
-            }
-        }
-        error_log('audit_log prune failed: ' . $e->getMessage());
-        return 0;
-    }
 }
 
 function prune_address_history(PDO $db, int $retentionDays): int
